Add files via upload
Number Guessing Game
This is a simple number-guessing game where the player tries to guess a randomly generated number between 1 and 100. The player has 5 attempts to guess the number. After each incorrect guess, the game provides a hint whether the guess should be higher or lower. The player can type "Quit" to exit the game early. After 5 incorrect guesses, the game reveals the correct number.